# springboot/target/generated-sources/openapi/src/main/java/com/nime/eqviewer/api/CodesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.nime.eqviewer.api;

import com.nime.eqviewer.dto.ErrorDto;
import com.nime.eqviewer.dto.SourceCodePairDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-11-30T16:25:57.029884200+08:00[Asia/Hong_Kong]")
@Validated
@Tag(name = "codes", description = "the codes API")
public interface CodesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /codes/{codesId} : Get one source_code_pairs
     *
     * @param codesId  (required)
     * @return A SourceCodePair (status code 200)
     *         or unexpected error (status code 200)
     */
    @Operation(
        operationId = "getSourceCodePair",
        summary = "Get one source_code_pairs",
        responses = {
            @ApiResponse(responseCode = "200", description = "A SourceCodePair", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  SourceCodePairDto.class))),
            @ApiResponse(responseCode = "200", description = "unexpected error", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ErrorDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/codes/{codesId}",
        produces = { "application/json" }
    )
    default ResponseEntity<SourceCodePairDto> getSourceCodePair(
        @Parameter(name = "codesId", description = "", required = true, schema = @Schema(description = "")) @PathVariable("codesId") Integer codesId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code2\" : { \"path\" : \"path\", \"id\" : 0 }, \"code1\" : { \"path\" : \"path\", \"id\" : 0 }, \"id\" : 6 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /codes : List all source_code_pairs
     *
     * @return A paged array of SourceCodePairs (status code 200)
     *         or unexpected error (status code 200)
     */
    @Operation(
        operationId = "listSourceCodePairs",
        summary = "List all source_code_pairs",
        tags = { "SourceCodePairs" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A paged array of SourceCodePairs", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  SourceCodePairDto.class))),
            @ApiResponse(responseCode = "200", description = "unexpected error", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ErrorDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/codes",
        produces = { "application/json" }
    )
    default ResponseEntity<List<SourceCodePairDto>> listSourceCodePairs(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code2\" : { \"path\" : \"path\", \"id\" : 0 }, \"code1\" : { \"path\" : \"path\", \"id\" : 0 }, \"id\" : 6 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /codes/{codesId} : Update a source_code_pairs
     *
     * @param codesId  (required)
     * @param sourceCodePairDto A JSON object containing code pair (optional)
     * @return Updated successfully (status code 200)
     *         or unexpected error (status code 200)
     */
    @Operation(
        operationId = "updateSourceCodePairs",
        summary = "Update a source_code_pairs",
        responses = {
            @ApiResponse(responseCode = "200", description = "Updated successfully", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  SourceCodePairDto.class))),
            @ApiResponse(responseCode = "200", description = "unexpected error", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ErrorDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/codes/{codesId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<SourceCodePairDto> updateSourceCodePairs(
        @Parameter(name = "codesId", description = "", required = true, schema = @Schema(description = "")) @PathVariable("codesId") Integer codesId,
        @Parameter(name = "SourceCodePairDto", description = "A JSON object containing code pair", schema = @Schema(description = "")) @Valid @RequestBody(required = false) SourceCodePairDto sourceCodePairDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code2\" : { \"path\" : \"path\", \"id\" : 0 }, \"code1\" : { \"path\" : \"path\", \"id\" : 0 }, \"id\" : 6 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}